# [performance] bundle bin, worker and index files