--- a/.eslintignore
+++ b/.eslintignore
@@ -1,2 +1,3 @@
 /dev/new-rule/templates/*
 /coverage
+/bin/ember-template-lint.js