--- a/lib/formatters/sarif.js
+++ b/lib/formatters/sarif.js
@@ -1,10 +1,14 @@
 import fs from 'node:fs';
-import { createRequire } from 'node:module';
 import url from 'node:url';
-
 import Linter from '../linter.js';
+import { findUpSync } from 'find-up';
 
-const require = createRequire(import.meta.url);
+function getVersion() {
+  const closestPackage = findUpSync('package.json');
+  // we can't use relative URL resolution because of bundling
+  const json = JSON.parse(fs.readFileSync(closestPackage));
+  return json.version;
+}
 
 function getResultLevel(message) {
   if (message.fatal || message.severity === 2) {
@@ -29,7 +33,7 @@ export default class SarifFormatter {
   }
 
   buildSarifLog(results) {
-    const version = require('../../package.json').version;
+    const version = getVersion();
     const sarifLog = {
       version: '2.1.0',
       $schema: 'http://json.schemastore.org/sarif-2.1.0-rtm.5',