--- a/lib/helpers/cli.js
+++ b/lib/helpers/cli.js
@@ -4,6 +4,7 @@ import ci from 'ci-info';
 import { globbySync } from 'globby';
 import isGlob from 'is-glob';
 import micromatch from 'micromatch';
+import { findUpSync } from 'find-up';
 import fs from 'node:fs';
 import path from 'node:path';
 import yargs from 'yargs';
@@ -25,8 +26,9 @@ function shouldIgnoreFileDueToConfig(filePath, config) {
 }
 
 function getVersion() {
-  const json = JSON.parse(fs.readFileSync(new URL('../../package.json', import.meta.url)));
-
+  const closestPackage = findUpSync('package.json');
+  // we can't use relative URL resolution because of bundling
+  const json = JSON.parse(fs.readFileSync(closestPackage));
   return json.version;
 }