--- a/lib/linter.js
+++ b/lib/linter.js
@@ -1,5 +1,5 @@
 import { parse, transform } from 'ember-template-recast';
-import * as contentTagUtils from 'content-tag-utils';
+import { Transformer } from 'content-tag-utils';
 
 import { canProcessFile, processFile } from './-private/file-path.js';
 import ModuleStatusCache from './-private/module-status-cache.js';
@@ -282,7 +282,7 @@ export default class Linter {
 
         let transformer;
         try {
-          transformer = new contentTagUtils.Transformer(source);
+          transformer = new Transformer(source);
         } catch (error) {
           let message = buildErrorMessage(options.filePath, error);
           results.push(message);
@@ -311,7 +311,7 @@ export default class Linter {
         return results;
       },
       fix: async (ruleNames, options, fileConfig) => {
-        let t = new contentTagUtils.Transformer(options.source);
+        let t = new Transformer(options.source);
 
         await t.asyncMap(async (source, coordinates) => {
           let currentSource = source;