--- a/lib/rules/require-lang-attribute.js
+++ b/lib/rules/require-lang-attribute.js
@@ -1,8 +1,6 @@
-import pkg from 'language-tags';
 import AstNodeInfo from '../helpers/ast-node-info.js';
 import Rule from './_base.js';
-
-const { check } = pkg;
+import codes from '../helpers/country-codes.js';
 
 const DEFAULT_CONFIG = {
   validateValues: true,
@@ -22,7 +20,9 @@ function hasValue(langAttrNode) {
 
 function hasValidValue(langAttrNode) {
   if (langAttrNode.value.type === 'TextNode') {
-    return check(langAttrNode.value.chars);
+    const value = langAttrNode.value.chars.toLowerCase();
+    const parts = value.split('-');
+    return parts.every((p) => codes.includes(p)) || codes.includes(value);
   } else {
     return langAttrNode.value !== undefined;
   }