--- a/lib/rules/require-lang-attribute.js
+++ b/lib/rules/require-lang-attribute.js
@@ -1,7 +1,8 @@
-import tags from 'language-tags';
-
+import pkg from 'language-tags';
 import AstNodeInfo from '../helpers/ast-node-info.js';
 import Rule from './_base.js';
+
+const { check } = pkg;
 
 const DEFAULT_CONFIG = {
   validateValues: true,
@@ -21,7 +22,7 @@ function hasValue(langAttrNode) {
 
 function hasValidValue(langAttrNode) {
   if (langAttrNode.value.type === 'TextNode') {
-    return tags.check(langAttrNode.value.chars);
+    return check(langAttrNode.value.chars);
   } else {
     return langAttrNode.value !== undefined;
   }