--- a/package.json
+++ b/package.json
@@ -51,6 +51,7 @@
     "test:watch": "vitest",
     "update": "npm-run-all update:*",
     "update:indices": "node ./scripts/update-indices.js",
+    "update:language-codes": "node ./scripts/update-readme.js",
     "update:readme": "node ./scripts/update-readme.js"
   },
   "dependencies": {
@@ -74,7 +75,6 @@
     "get-stdin": "^9.0.0",
     "globby": "^14.0.1",
     "is-glob": "^4.0.3",
-    "language-tags": "^1.0.8",
     "micromatch": "^4.0.5",
     "resolve": "^1.22.3",
     "v8-compile-cache": "^2.3.0",