Metrics:

Test suite duration (master):
```
Test Files  156 passed (156)
      Tests  8860 passed | 5 skipped (8865)
   Start at  13:06:30
   Duration  55.81s (transform 1.29s, setup 0ms, collect 77.10s, tests 120.25s, environment 24ms, prepare 11.04s)
```

~Test suite duration (bundled (esbuild)):~
```
 Test Files  159 passed (159)
      Tests  8979 passed | 5 skipped (8984)
   Start at  17:00:47
   Duration  41.58s (transform 1.27s, setup 0ms, collect 74.95s, tests 85.65s, environment 21ms, prepare 11.44s)
```

Tes suite duration (bundled (rolldown))
```
 Test Files  159 passed (159)
      Tests  8979 passed | 5 skipped (8984)
   Start at  19:29:11
   Duration  33.75s (transform 1.21s, setup 0ms, collect 70.59s, tests 69.77s, environment 24ms, prepare 10.45s)
```

Improvement: `120.25s` -> `69.77s`

Note, it's likely an startup improvement, not overall linting x2 improvement.

Minified bin bundle size: `1.6 mb`

Huuuge dependency: `language-tags`, to save `800kb` of bundle size we have builtin country code download & resolution logic, it's not 100% spec accurate, but, 800kb is 800kb.

Some dependencies is not really bundlabe, because of error like this: https://github.com/evanw/esbuild/issues/3324, https://github.com/evanw/esbuild/issues/1921

Note:
~We need to pin dependencies to exact versions to have same behaviour as we have if import package directly.~ (now we bundling public api scope, so it's should be safe)