@NullVoxPopuli 2025-05-16
@SergeAstapov how'd it go?