@SergeAstapov 2025-05-17
@lifeart @NullVoxPopuli sorry for delay, on my project I see following numbers:

```
~/Code/[my app]/apps/[my app]
time /Users/sastapov/Code/tmp/ember-template-lint/bin/ember-template-lint.js .
Linting 2431 Total Files with TemplateLint
        .js: 1345
        .ts: 579
        .hbs: 284
        .gts: 203
        .gjs: 2
        .scenario.js: 18


/Users/sastapov/Code/tmp/ember-template-lint/bin/ember-template-lint.js .  25.94s user 1.98s system 413% cpu 6.756 total


~/Code/[my app]/apps/[my app]
time pnpm run lint:hbs

> [my app]@1.0.0 lint:hbs /Users/sastapov/Code/[my app]/apps/[my app]
> ember-template-lint .

Linting 2431 Total Files with TemplateLint
        .js: 1345
        .ts: 579
        .hbs: 284
        .gts: 203
        .gjs: 2
        .scenario.js: 18


pnpm run lint:hbs  47.10s user 2.80s system 461% cpu 10.811 total
```

so I do see about 2x speed improvement! 🤯