@lifeart 2025-05-17
@SergeAstapov wow! Impressive numbers! 
Wondering, if results is better/worse if we change `MIN_FILES_TO_USE_WORKERS` from `100` to `150` or `200`.